pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ddbj-search-converter"
authors = [{ name = "Bioinformatics and DDBJ Center" }]
readme = { file = "README.md", content-type = "text/markdown" }
version = "0.1.0"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "duckdb",
    "elasticsearch>=8,<9",
    "httpx",
    "lxml",
    "psycopg2-binary",
    "pydantic",
]

[project.urls]
Homepage = "https://github.com/ddbj/ddbj-search-converter"
Documentation = "https://github.com/ddbj/ddbj-search-converter/blob/main/README.md"
Repository = "https://github.com/ddbj/ddbj-search-converter.git"

[project.optional-dependencies]
tests = [
    "hypothesis",
    "lxml-stubs",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "ruff",
]

[project.scripts]
check_external_resources = "ddbj_search_converter.cli.check_external_resources:main"
prepare_biosample_xml = "ddbj_search_converter.cli.prepare_biosample_xml:main"
prepare_bioproject_xml = "ddbj_search_converter.cli.prepare_bioproject_xml:main"
build_sra_and_dra_accessions_db = "ddbj_search_converter.sra_accessions_tab:main"
init_dblink_db = "ddbj_search_converter.cli.init_dblink_db:main"
create_dblink_assembly_and_master_relations = "ddbj_search_converter.dblink.assembly_and_master:main"
create_dblink_bp_relations = "ddbj_search_converter.dblink.bioproject:main"
create_dblink_bp_bs_relations = "ddbj_search_converter.dblink.bp_bs:main"
create_dblink_gea_relations = "ddbj_search_converter.dblink.gea:main"
create_dblink_metabobank_relations = "ddbj_search_converter.dblink.metabobank:main"
create_dblink_jga_relations = "ddbj_search_converter.dblink.jga:main"
create_dblink_sra_internal_relations = "ddbj_search_converter.dblink.sra_internal:main"
finalize_dblink_db = "ddbj_search_converter.cli.finalize_dblink_db:main"
dump_dblink_files = "ddbj_search_converter.cli.dump_dblink_files:main"
sync_ncbi_tar = "ddbj_search_converter.cli.sync_ncbi_tar:main"
sync_dra_tar = "ddbj_search_converter.cli.sync_dra_tar:main"
build_bp_bs_date_cache = "ddbj_search_converter.cli.build_bp_bs_date_cache:main"
generate_jga_jsonl = "ddbj_search_converter.jsonl.jga:main"
generate_bp_jsonl = "ddbj_search_converter.jsonl.bp:main"
generate_bs_jsonl = "ddbj_search_converter.jsonl.bs:main"
generate_sra_jsonl = "ddbj_search_converter.jsonl.sra:main"
regenerate_jsonl = "ddbj_search_converter.jsonl.regenerate:main"
es_create_index = "ddbj_search_converter.cli.es:main_create_index"
es_delete_index = "ddbj_search_converter.cli.es:main_delete_index"
es_bulk_insert = "ddbj_search_converter.cli.es:main_bulk_insert"
es_list_indexes = "ddbj_search_converter.cli.es:main_list_indexes"
es_health_check = "ddbj_search_converter.cli.es:main_health_check"
es_delete_blacklist = "ddbj_search_converter.cli.es:main_delete_blacklist"
es_snapshot = "ddbj_search_converter.cli.es_snapshot:main"
show_log_summary = "ddbj_search_converter.cli.debug.show_log_summary:main"
show_log = "ddbj_search_converter.cli.debug.show_log:main"
show_dblink_counts = "ddbj_search_converter.cli.debug.show_dblink_counts:main"


[tool.pytest.ini_options]
addopts = "--cov=ddbj_search_converter --cov-report=html:tests/htmlcov"
testpaths = ["tests/py_tests"]

[tool.mypy]
files = ["./ddbj_search_converter/**/*.py", "./tests/py_tests/**/*.py"]
namespace_packages = true
explicit_package_bases = true
follow_imports = "silent"
strict = true
python_version = "3.10"
warn_unreachable = true

[[tool.mypy.overrides]]
module = "psycopg2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.py_tests.*"
disallow_untyped_defs = false
warn_unreachable = false

[[tool.mypy.overrides]]
module = "py_tests.*"
disallow_untyped_defs = false
warn_unreachable = false
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # pydocstyle (docstrings)
    "ANN",     # flake8-annotations (strict は mypy に任せる)
    "ERA",     # eradicate (commented-out code)
    "FIX",     # fixme comments
    "TD",      # todo comments
    "S",       # flake8-bandit (security)
    "INP",     # implicit namespace packages
    "T20",     # flake8-print
    "FBT",     # flake8-boolean-trap
    "EM",      # flake8-errmsg
    "TRY002",  # create own exception
    "TRY003",  # long exception messages
    "TRY300",  # consider moving into else block
    "TRY301",  # abstract raise to inner function
    "TRY004",  # prefer TypeError
    "PLC0415", # import-outside-toplevel (intentional lazy imports)
    "COM812",  # trailing comma (conflicts with formatter)
    "ISC001",  # implicit string concat (conflicts with formatter)
    "PLR0911", # too many return statements
    "PLR0912", # too many branches
    "PLR0913", # too many arguments
    "PLR0915", # too many statements
    "PLR2004", # magic value comparison
    "C901",    # too complex
    "BLE001",  # broad exception catching (intentional in pipeline code)
    "N815",    # mixed case variable in class scope (Pydantic field names match JSON API)
    "RUF002",  # ambiguous unicode in docstring (Japanese)
    "RUF003",  # ambiguous unicode in comment (Japanese)
    "RUF012",  # mutable class variable without ClassVar
    "DTZ",     # datetime timezone (naive datetimes intentional for DB)
    "PERF203", # try-except in loop (micro-optimization)
    "TC",      # type checking imports
    "PYI034",  # __enter__ should return Self
    "SIM108",  # ternary operator (readability preference)
    "RET504",  # unnecessary assignment before return
    "SIM115",  # context manager for open (false positive for resource management classes)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",     # unused arguments (fixtures)
    "PLR2004", # magic values in tests
    "SLF001",  # private member access in tests
    "PT011",   # pytest.raises too broad (intentional for testing error handling)
    "B017",    # assert blind exception (intentional for testing error handling)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = ["ddbj_search_converter"]
include = ["ddbj_search_converter/py.typed"]
